[back] dto 폴더추가

+++ server/GoalForYou/src/main/java/com/example/GoalForYou/domain/Badge.java
@@ -10,8 +10,10 @@ public class Badge {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer badgeId;
+    private Long badgeId;
+
     private String badgeName;
+
     private String badgeImage;
 
 }
